requirements.txt
azure-core==1.23.1
azure-cosmos==4.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
